feature - added logic to avoid delete topics with child

+++ src/modules/topic/application/useCases/deleteTopic.ts
@@ -15,6 +15,11 @@ export class DeleteTopic {
             throw new Error(`Topic with id ${id} not found`);
         }
         
+        const childTopics = await this.topicRepository.findByParentId(id);
+        if (childTopics.length > 0) {
+            throw new Error(`Cannot delete topic with id ${id} because it has child topics. Delete all child topics first.`);
+        }
+        
         await this.resourceCommandHandler.deleteResourcesByTopicId(id);
         
         await this.topicRepository.delete(id);
